requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
pypdf
python-dotenv
tiktoken
langchain-text-splitters
lxml
html2text
youtube-transcript-api
pytube
yt_dlp
streamlit
python-magic
unstructured
python-pptx
docx2txt
moviepy==1.0.3
google-generativeai